The ObservableCache now uses replay and autoConnect instead of share and subscribeWith to enable proper/full disposal of the Observable chain
More info on the disposing the whole chain part: calling dispose on the observable returned by the subscribeWith call only disposes from the subscribeWith part onward,
the underlying chain is kept intact and there is no good way to get rid of it. With the replay and autoConnect version the functionality stays the same, and the
underlying chain may be disposed by calling dispose on the Disposables gathered by the second argument in the autoConnect call.